Tue, Jun 21, 2022, 12:31 AM +00:00

+++ pset2/readability/readability.c
@@ -64,7 +64,7 @@ void grade(string text)
 {
     int L = 100 * count_letters(text) / count_words(text); //L is the average number of letters per 100 words in the text
     int S = 100 * count_sentences(text) / count_words(text); //S is the average number of sentences per 100 words in the text
-    float index = 0.0588 * L - 0.296 * S - 15.8;
+    int index = 0.0588 * L - 0.296 * S - 15.8;
     if (round(index) < 1)
     {
         printf("Before Grade 1\n");
@@ -74,7 +74,7 @@ void grade(string text)
         printf("Grade 16+\n");
     }
     else{
-        printf("Grade %.2f\n", index);
+        printf("Grade %i\n", index);
     }
 }
